Refactor log.py to avoid creating duplicate loggers
Before this change, when get_logger() is called outside of a function in
the consuming application, a new logger would get created and added to
a daisy chained list of loggers. This change prevents that from
happening no matter where get_logger() gets called.

The config was also changed to a dict for easier expansion in the
future.